Refactor: Extract FreeCellGameBase for core game logic
Introduce FreeCellGameBase to encapsulate core FreeCell logic (move validation, auto-move, sequence detection) without undo/serialization. Move SourceType and CardSelection to the base class for reuse. Refactor FreeCellGameService to inherit from FreeCellGameBase, removing duplicated logic and implementing undo via OnBeforeMove(). This separation improves maintainability, code reuse, and testability.
