+++ Client/Games/Cards/Services/FreeCellGameService.cs
@@ -1,26 +1,7 @@
 using Client.Games.Cards.Models;
-using System.Text;
 using System.Text.Json;
 
 namespace Client.Games.Cards.Services;
-
-/// <summary>
-/// Source of a card selection or move target
-/// </summary>
-public enum SourceType
-{
-    FreeCell = 0,
-    Tableau = 1,
-    Foundation = 2
-}
-
-/// <summary>
-/// Represents a card selection with source location
-/// </summary>
-/// <param name="SourceType">The type of source (FreeCell, Tableau, or Foundation)</param>
-/// <param name="SourceIndex">Index within the source (column index for Tableau, cell index for FreeCell, pile index for Foundation)</param>
-/// <param name="CardIndex">Index of the card within the source (for Tableau stacks)</param>
-public readonly record struct CardSelection(SourceType SourceType, int SourceIndex, int CardIndex);
 
 /// <summary>
 /// DTO for serializing FreeCell game state
@@ -36,22 +17,12 @@ public class FreeCellGameState
 }
 
 /// <summary>
-/// Service for managing FreeCell game state
+/// Full-featured FreeCell game service with undo support and serialization.
+/// Inherits core game logic from FreeCellGameBase.
 /// </summary>
-public class FreeCellGameService
+public class FreeCellGameService : FreeCellGameBase
 {
     private readonly Random _random;
-
-    // Game identification
-    public int GameId { get; private set; }
-
-    // Game state
-    public List<List<Card>> Tableau { get; private set; } = new(); // 8 columns
-    public List<Card?> FreeCells { get; private set; } = new(); // 4 free cells
-    public List<List<Card>> Foundations { get; private set; } = new(); // 4 foundation piles
-
-    // Selection state
-    public CardSelection? Selection { get; set; }
 
     // Undo support
     private readonly Stack<GameSnapshot> _undoStack = new();
@@ -62,26 +33,8 @@ public class FreeCellGameService
         int MoveCount
     );
 
-    public int MoveCount { get; private set; }
-    public bool IsGameWon => Foundations.All(f => f.Count == 13);
     public bool CanUndo => _undoStack.Count > 0;
     public int UndoCount => _undoStack.Count;
-
-    /// <summary>
-    /// Checks if the game is in a stalemate (no valid moves available).
-    /// This is a losing condition - the player cannot make any moves.
-    /// </summary>
-    public bool IsStalemate
-    {
-        get
-        {
-            // Not a stalemate if game is already won
-            if (IsGameWon) return false;
-
-            // Check all possible moves
-            return !HasAnyValidMove();
-        }
-    }
 
     public FreeCellGameService(Random? random = null)
     {
@@ -398,48 +351,6 @@ public class FreeCellGameService
     }
 
     /// <summary>
-    /// Gets the number of empty free cells
-    /// </summary>
-    public int EmptyFreeCellCount => FreeCells.Count(c => c == null);
-
-    /// <summary>
-    /// Gets the number of empty tableau columns
-    /// </summary>
-    public int EmptyTableauCount => Tableau.Count(col => col.Count == 0);
-
-    /// <summary>
-    /// Calculates the maximum number of cards that can be moved as a stack
-    /// Formula: (1 + empty free cells) * 2^(empty columns)
-    /// </summary>
-    public int MaxMovableCards
-    {
-        get
-        {
-            int emptyFreeCells = EmptyFreeCellCount;
-            int emptyColumns = EmptyTableauCount;
-            return (1 + emptyFreeCells) * (int)Math.Pow(2, emptyColumns);
-        }
-    }
-
-    /// <summary>
-    /// Selects a card or stack of cards
-    /// </summary>
-    public void Select(SourceType sourceType, int sourceIndex, int cardIndex = -1)
-    {
-        Selection = new CardSelection(sourceType, sourceIndex, cardIndex);
-    }
-
-    // (Removed integer overload — callers updated to use SourceType enum)
-
-    /// <summary>
-    /// Clears the current selection
-    /// </summary>
-    public void ClearSelection()
-    {
-        Selection = null;
-    }
-
-    /// <summary>
     /// Captures the current game state for undo
     /// </summary>
     private GameSnapshot CaptureSnapshot()
@@ -450,6 +361,14 @@ public class FreeCellGameService
             Foundations.Select(f => f.Select(c => new Card(c.Suit, c.Rank, c.IsFaceUp)).ToList()).ToList(),
             MoveCount
         );
+    }
+
+    /// <summary>
+    /// Override to capture snapshot before each move for undo support
+    /// </summary>
+    protected override void OnBeforeMove()
+    {
+        _undoStack.Push(CaptureSnapshot());
     }
 
     /// <summary>
@@ -474,531 +393,6 @@ public class FreeCellGameService
         var snapshot = _undoStack.Pop();
         RestoreSnapshot(snapshot);
         return true;
-    }
-
-    /// <summary>
-    /// Attempts to move selected cards to a target
-    /// </summary>
-    public bool TryMove(SourceType targetType, int targetIndex)
-    {
-        if (Selection == null) return false;
-
-        var (sourceType, sourceIndex, cardIndex) = (Selection.Value.SourceType, Selection.Value.SourceIndex, Selection.Value.CardIndex);
-        List<Card> cardsToMove = new();
-
-        // Get cards to move
-        switch (sourceType)
-        {
-            case SourceType.FreeCell: // FreeCell
-                if (sourceIndex < 0 || sourceIndex >= 4) return false;
-                var freeCard = FreeCells[sourceIndex];
-                if (freeCard == null) return false;
-                cardsToMove.Add(freeCard);
-                break;
-            case SourceType.Tableau: // Tableau
-                if (sourceIndex < 0 || sourceIndex >= Tableau.Count) return false;
-                var column = Tableau[sourceIndex];
-                if (cardIndex < 0 || cardIndex >= column.Count) return false;
-
-                // Validate the stack is properly ordered (descending, alternating colors)
-                if (!IsValidTableauStack(column, cardIndex)) return false;
-
-                // Check if we can move this many cards
-                int cardCount = column.Count - cardIndex;
-                int maxMovable = CalculateMaxMovableCards(targetType, targetIndex);
-                if (cardCount > maxMovable) return false;
-
-                // Now get the cards to move (after validation)
-                cardsToMove = column.GetRange(cardIndex, cardCount);
-                break;
-            case SourceType.Foundation: // Foundation
-                if (sourceIndex < 0 || sourceIndex >= Foundations.Count) return false;
-                if (Foundations[sourceIndex].Count == 0) return false;
-                cardsToMove.Add(Foundations[sourceIndex][^1]);
-                break;
-            default:
-                return false;
-        }
-
-        if (cardsToMove.Count == 0) return false;
-
-        // Try to place on target
-        bool success = false;
-        switch (targetType)
-        {
-            case SourceType.FreeCell: // FreeCell
-                if (targetIndex >= 0 && targetIndex < 4 && cardsToMove.Count == 1)
-                {
-                    if (FreeCells[targetIndex] == null)
-                    {
-                        // Save state before move
-                        _undoStack.Push(CaptureSnapshot());
-
-                        FreeCells[targetIndex] = cardsToMove[0];
-                        success = true;
-                    }
-                }
-                break;
-            case SourceType.Tableau: // Tableau
-                if (targetIndex >= 0 && targetIndex < Tableau.Count)
-                {
-                    if (CanPlaceOnTableau(cardsToMove[0], Tableau[targetIndex]))
-                    {
-                        // Save state before move
-                        _undoStack.Push(CaptureSnapshot());
-
-                        Tableau[targetIndex].AddRange(cardsToMove);
-                        success = true;
-                    }
-                }
-                break;
-            case SourceType.Foundation: // Foundation
-                if (targetIndex >= 0 && targetIndex < Foundations.Count && cardsToMove.Count == 1)
-                {
-                    if (CanPlaceOnFoundation(cardsToMove[0], Foundations[targetIndex]))
-                    {
-                        // Save state before move
-                        _undoStack.Push(CaptureSnapshot());
-
-                        Foundations[targetIndex].Add(cardsToMove[0]);
-                        success = true;
-                    }
-                }
-                break;
-        }
-
-        if (success)
-        {
-            // Remove from source
-            switch (sourceType)
-            {
-                case SourceType.FreeCell: // FreeCell
-                    FreeCells[sourceIndex] = null;
-                    break;
-                case SourceType.Tableau: // Tableau
-                    Tableau[sourceIndex].RemoveRange(cardIndex, cardsToMove.Count);
-                    break;
-                case SourceType.Foundation: // Foundation
-                    Foundations[sourceIndex].RemoveAt(Foundations[sourceIndex].Count - 1);
-                    break;
-            }
-
-            MoveCount++;
-            Selection = null;
-        }
-
-        return success;
-    }
-
-    /// <summary>
-    /// Attempts to auto-move a card to foundation
-    /// </summary>
-    public bool TryAutoMoveToFoundation(SourceType sourceType, int sourceIndex, int cardIndex = -1)
-    {
-        Card? card = null;
-
-        switch (sourceType)
-        {
-            case SourceType.FreeCell: // FreeCell
-                if (sourceIndex >= 0 && sourceIndex < 4)
-                    card = FreeCells[sourceIndex];
-                break;
-            case SourceType.Tableau: // Tableau
-                if (sourceIndex >= 0 && sourceIndex < Tableau.Count && Tableau[sourceIndex].Count > 0)
-                {
-                    card = Tableau[sourceIndex][^1];
-                    cardIndex = Tableau[sourceIndex].Count - 1;
-                }
-                break;
-        }
-
-        if (card == null) return false;
-
-        // Find appropriate foundation
-        for (int i = 0; i < Foundations.Count; i++)
-        {
-            if (CanPlaceOnFoundation(card, Foundations[i]))
-            {
-                Selection = new CardSelection(sourceType, sourceIndex, cardIndex);
-                return TryMove(SourceType.Foundation, i);
-            }
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Auto-moves all possible cards to foundations
-    /// </summary>
-    public int AutoMoveToFoundations()
-    {
-        int movesMade = 0;
-        bool madeMove;
-
-        do
-        {
-            madeMove = false;
-
-            // Try free cells
-            for (int i = 0; i < 4; i++)
-            {
-                if (FreeCells[i] != null && TryAutoMoveToFoundation(SourceType.FreeCell, i))
-                {
-                    madeMove = true;
-                    movesMade++;
-                }
-            }
-
-            // Try tableau columns
-            for (int i = 0; i < 8; i++)
-            {
-                if (Tableau[i].Count > 0 && TryAutoMoveToFoundation(SourceType.Tableau, i))
-                {
-                    madeMove = true;
-                    movesMade++;
-                }
-            }
-        } while (madeMove);
-
-        return movesMade;
-    }
-
-    /// <summary>
-    /// Checks if the game is in a trivially winnable state.
-    /// A game is trivially winnable when all cards in tableau are in descending sequences
-    /// and can be moved to foundations without any complex moves.
-    /// </summary>
-    public (bool isWinnable, string reason) IsTriviallyWinnableWithReason()
-    {
-        // If already won, return false (nothing to do)
-        if (IsGameWon) return (false, "Game already won");
-
-        // Check if all tableau columns are in valid descending sequences from top to bottom
-        for (int colIdx = 0; colIdx < Tableau.Count; colIdx++)
-        {
-            var column = Tableau[colIdx];
-            if (column.Count <= 1) continue;
-
-            // Check if the entire column is a valid descending sequence
-            for (int i = 0; i < column.Count - 1; i++)
-            {
-                var current = column[i];
-                var next = column[i + 1];
-
-                // Cards must be in descending order (same rank is OK - they go to different foundations)
-                if ((int)current.Rank < (int)next.Rank)
-                {
-                    return (false, $"Column {colIdx + 1}: {current} (rank {(int)current.Rank}) is not >= {next} (rank {(int)next.Rank}) at position {i}");
-                }
-            }
-        }
-
-        // Check free cells - all cards in free cells must be playable to foundations eventually
-        // For simplicity, we consider it trivially winnable if tableau is all in order
-        // The free cell cards will be moved when their turn comes
-
-        return (true, "All columns in descending order");
-    }
-
-    /// <summary>
-    /// Checks if the game is in a trivially winnable state.
-    /// A game is trivially winnable when all cards in tableau are in descending sequences
-    /// and can be moved to foundations without any complex moves.
-    /// </summary>
-    public bool IsTriviallyWinnable()
-    {
-        return IsTriviallyWinnableWithReason().isWinnable;
-    }
-
-    /// <summary>
-    /// Gets the next card that can be moved to a foundation.
-    /// Returns the CardSelection or null if none found.
-    /// </summary>
-    public CardSelection? GetNextFoundationMove()
-    {
-        // Check free cells first
-        for (int i = 0; i < 4; i++)
-        {
-            var card = FreeCells[i];
-            if (card != null && CanMoveToAnyFoundation(card))
-            {
-                return new CardSelection(SourceType.FreeCell, i, 0);
-            }
-        }
-
-        // Check tableau columns (top card only)
-        for (int col = 0; col < 8; col++)
-        {
-            if (Tableau[col].Count > 0)
-            {
-                var card = Tableau[col][^1];
-                if (CanMoveToAnyFoundation(card))
-                {
-                    return new CardSelection(SourceType.Tableau, col, Tableau[col].Count - 1);
-                }
-            }
-        }
-
-        return null;
-    }
-
-    /// <summary>
-    /// Checks if a card can be moved to any foundation pile
-    /// </summary>
-    public bool CanMoveToAnyFoundation(Card card)
-    {
-        for (int i = 0; i < Foundations.Count; i++)
-        {
-            if (CanPlaceOnFoundation(card, Foundations[i]))
-            {
-                return true;
-            }
-        }
-        return false;
-    }
-    public bool CanMoveTableauToTableau(int sourceCol, int targetCol, int length)
-    {
-        if (sourceCol < 0 || sourceCol >= Tableau.Count || targetCol < 0 || targetCol >= Tableau.Count) return false;
-        if (Tableau[sourceCol].Count == 0) return false;
-        var cardToMove = Tableau[sourceCol][^length];
-        return CanPlaceOnTableau(cardToMove, Tableau[targetCol]);
-    }
-    /// <summary>
-    /// Performs one step of auto-solve by moving a card to foundation.
-    /// Returns info about the move made, or null if no move possible.
-    /// </summary>
-    public (SourceType sourceType, int sourceIndex, Card card)? AutoMoveStep()
-    {
-        var nextMove = GetNextFoundationMove();
-        if (nextMove == null) return null;
-
-        var (sourceType, sourceIndex, cardIndex) = (nextMove.Value.SourceType, nextMove.Value.SourceIndex, nextMove.Value.CardIndex);
-
-        Card? card = sourceType switch
-        {
-            SourceType.FreeCell => FreeCells[sourceIndex],
-            SourceType.Tableau => Tableau[sourceIndex].Count > 0 ? Tableau[sourceIndex][^1] : null,
-            _ => null
-        };
-
-        if (card == null) return null;
-
-        // Make a copy before move
-        var movedCard = new Card(card.Suit, card.Rank, true);
-
-        Selection = nextMove;
-        if (TryAutoMoveToFoundation(sourceType, sourceIndex, cardIndex))
-        {
-            return (sourceType, sourceIndex, movedCard);
-        }
-
-        return null;
-    }
-
-    /// <summary>
-    /// Checks if selecting from a specific card index in a tableau column forms a valid sequence
-    /// </summary>
-    public bool IsValidTableauSequence(int columnIndex, int cardIndex)
-    {
-        if (columnIndex < 0 || columnIndex >= Tableau.Count) return false;
-        var column = Tableau[columnIndex];
-        if (cardIndex < 0 || cardIndex >= column.Count) return false;
-
-        return IsValidTableauStack(column, cardIndex);
-    }
-
-    /// <summary>
-    /// Validates that a stack of cards is properly ordered for tableau movement.
-    /// Uses indices to avoid List allocations.
-    /// </summary>
-    /// <param name="column">The tableau column</param>
-    /// <param name="startIndex">Starting index in the column</param>
-    /// <returns>True if cards from startIndex to end form a valid descending alternating-color sequence</returns>
-    private bool IsValidTableauStack(List<Card> column, int startIndex)
-    {
-        int count = column.Count - startIndex;
-        if (count <= 1) return true;
-
-        for (int i = startIndex; i < column.Count - 1; i++)
-        {
-            var current = column[i];
-            var next = column[i + 1];
-
-            // Must be descending rank and alternating colors
-            if ((int)current.Rank != (int)next.Rank + 1) return false;
-            if (current.IsRed == next.IsRed) return false;
-        }
-
-        return true;
-    }
-
-    /// <summary>
-    /// Gets the starting index of the largest valid tableau sequence at the bottom of a column.
-    /// A valid sequence is descending rank with alternating colors.
-    /// </summary>
-    /// <param name="columnIndex">The tableau column index (0-7)</param>
-    /// <returns>The starting index of the sequence, or -1 if column is empty or invalid</returns>
-    public int GetBottomSequenceStartIndex(int columnIndex)
-    {
-        if (columnIndex < 0 || columnIndex >= Tableau.Count) return -1;
-        var column = Tableau[columnIndex];
-        if (column.Count == 0) return -1;
-        if (column.Count == 1) return 0;
-
-        // Start from the second-to-last card and work upward
-        int sequenceStart = column.Count - 1;
-
-        for (int i = column.Count - 2; i >= 0; i--)
-        {
-            var current = column[i];
-            var next = column[i + 1];
-
-            // Check if current and next form a valid sequence (descending, alternating colors)
-            if ((int)current.Rank == (int)next.Rank + 1 && current.IsRed != next.IsRed)
-            {
-                sequenceStart = i;
-            }
-            else
-            {
-                break; // Sequence is broken
-            }
-        }
-
-        return sequenceStart;
-    }
-
-    /// <summary>
-    /// Gets the length of the largest valid tableau sequence at the bottom of a column.
-    /// </summary>
-    /// <param name="columnIndex">The tableau column index (0-7)</param>
-    /// <returns>The number of cards in the sequence, or 0 if column is empty or invalid</returns>
-    public int GetBottomSequenceLength(int columnIndex)
-    {
-        int startIndex = GetBottomSequenceStartIndex(columnIndex);
-        if (startIndex < 0) return 0;
-        return Tableau[columnIndex].Count - startIndex;
-    }
-
-    public List<int> GetBottomSequenceLengths()
-    {
-        List<int> lengths = new();
-        for (int col = 0; col < Tableau.Count; col++)
-        {
-            lengths.Add(GetBottomSequenceLength(col));
-        }
-        return lengths;
-    }
-
-    /// <summary>
-    /// Calculates max movable cards considering the target
-    /// </summary>
-    private int CalculateMaxMovableCards(SourceType targetType, int targetIndex)
-    {
-        int emptyFreeCells = EmptyFreeCellCount;
-        int emptyColumns = EmptyTableauCount;
-
-        // If moving to an empty column, we have one fewer empty column to use
-        if (targetType == SourceType.Tableau && targetIndex >= 0 && targetIndex < Tableau.Count && Tableau[targetIndex].Count == 0)
-        {
-            emptyColumns = Math.Max(0, emptyColumns - 1);
-        }
-
-        return (1 + emptyFreeCells) * (int)Math.Pow(2, emptyColumns);
-    }
-
-    // (Removed integer overload — callers updated to use SourceType enum)
-
-    /// <summary>
-    /// Checks if a card can be placed on a tableau column
-    /// </summary>
-    private bool CanPlaceOnTableau(Card card, List<Card> column)
-    {
-        if (column.Count == 0)
-        {
-            // Any card can go on an empty column in FreeCell
-            return true;
-        }
-
-        var topCard = column[^1];
-        // Must be opposite color and one rank lower
-        return card.IsRed != topCard.IsRed && (int)card.Rank == (int)topCard.Rank - 1;
-    }
-
-    /// <summary>
-    /// Checks if a card can be placed on a foundation pile
-    /// </summary>
-    private bool CanPlaceOnFoundation(Card card, List<Card> foundation)
-    {
-        if (foundation.Count == 0)
-        {
-            // Only Aces can start a foundation
-            return card.Rank == Rank.Ace;
-        }
-
-        var topCard = foundation[^1];
-        // Must be same suit and one rank higher
-        return card.Suit == topCard.Suit && (int)card.Rank == (int)topCard.Rank + 1;
-    }
-
-    /// <summary>
-    /// Checks if there is at least one valid move available.
-    /// </summary>
-    private bool HasAnyValidMove()
-    {
-        // Check moves from free cells
-        for (int i = 0; i < 4; i++)
-        {
-            var card = FreeCells[i];
-            if (card == null) continue;
-
-            // Can move to foundation?
-            if (CanMoveToAnyFoundation(card)) return true;
-
-            // Can move to any tableau column?
-            for (int col = 0; col < 8; col++)
-            {
-                if (CanPlaceOnTableau(card, Tableau[col])) return true;
-            }
-        }
-
-        // Check moves from tableau
-        for (int col = 0; col < 8; col++)
-        {
-            var column = Tableau[col];
-            if (column.Count == 0) continue;
-
-            // Top card can move to foundation?
-            var topCard = column[^1];
-            if (CanMoveToAnyFoundation(topCard)) return true;
-
-            // Check all valid sequences starting from this column
-            for (int cardIdx = column.Count - 1; cardIdx >= 0; cardIdx--)
-            {
-                // Is this the start of a valid sequence?
-                if (!IsValidTableauStack(column, cardIdx)) continue;
-
-                var leadCard = column[cardIdx];
-                int stackSize = column.Count - cardIdx;
-
-                // Single card can move to empty free cell?
-                if (stackSize == 1 && EmptyFreeCellCount > 0) return true;
-
-                // Can move to any other tableau column?
-                for (int targetCol = 0; targetCol < 8; targetCol++)
-                {
-                    if (targetCol == col) continue;
-
-                    // Check if can place and have enough capacity
-                    if (CanPlaceOnTableau(leadCard, Tableau[targetCol]))
-                    {
-                        int maxMovable = CalculateMaxMovableCards(SourceType.Tableau, targetCol);
-                        if (stackSize <= maxMovable) return true;
-                    }
-                }
-            }
-        }
-
-        return false;
     }
 
     #region Serialization
